Fix panic if todo is not part of a project
According to https://docs.gitlab.com/ee/user/todos.html there are todos that are not part of a project. For example member access request for groups are only tied to the group, not a project. This PR fixes the NPE by checking if a projects is there first

There are no tests that cover the outgoing API requests and hence the changes in this PR are not covered by tests